+++ frameworks/native/hilog_input_socket_client.cpp
@@ -38,7 +38,7 @@ int HilogInputSocketClient::WriteLogMessage(HilogMsg *header, const char *tag, u
 
     struct timeval tv = {0};
     gettimeofday(&tv, nullptr);
-    header->tv_sec = tv.tv_sec;
+    header->tv_sec = static_cast<uint32_t>(tv.tv_sec);
     header->tv_nsec = static_cast<uint32_t>(tv.tv_usec * 1000);     // 1000 : usec convert to nsec
     header->len = sizeof(HilogMsg) + tagLen + fmtLen;
     header->tag_len = tagLen;
